Some small changes

+++ src/components/projects/ProjectDetails.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="container">
+  <div class="container mt-5">
     <div class="row">
       <div class="col-lg-8 mx-auto">
         <div class="card">
@@ -7,10 +7,19 @@
             <h2 class="text-center">{{ project.name }}</h2>
           </div>
           <div class="card-body">
-            <p class="text-justify">
-              <i class="fa fa-user"></i>
-              {{ project.description }}
-            </p>
+            <label class="form-label">Description</label>
+            <textarea
+              cols="30"
+              rows="5"
+              class="form-control"
+              v-model="project.description"
+              :readonly="!showTextarea"
+            >
+            </textarea>
+            <hr />
+            <button class="btn btn-danger" @click="deleteProject">
+              Delete
+            </button>
           </div>
         </div>
       </div>
@@ -19,7 +28,7 @@
 </template>
 
 <script>
-import { onMounted, reactive } from "vue";
+import { onMounted, reactive, ref } from "vue";
 import { useRoute } from "vue-router";
 import { db } from "@/firebase";
 import { deleteDoc, doc, getDoc } from "firebase/firestore";
@@ -30,21 +39,30 @@ export default {
     const projectId = route.params.projectId;
     const projectRef = doc(db, "projects", projectId);
     let project = reactive({});
+    const showTextarea = ref(false);
+
+    function makeDescriptionEditable() {
+      showTextarea.value = !showTextarea.value;
+    }
 
     onMounted(async () => {
       const projectSnap = await getDoc(projectRef);
       if (projectSnap.exists()) {
-        Object.assign(project, projectSnap.data());        
+        Object.assign(project, projectSnap.data());
       }
     });
 
     function deleteProject() {
-      deleteDoc(projectRef);
+      if (confirm("Are you sure you want to delete the project")) {
+        deleteDoc(projectRef);
+      }
     }
 
     return {
       project,
-      deleteProject
+      showTextarea,
+      deleteProject,
+      makeDescriptionEditable,
     };
   },
 };
